move evaluationstack out of detail
