+++ libs/server-sdk/src/client_impl.cpp
@@ -4,7 +4,7 @@
 #include "data_systems/background_sync/background_sync_system.hpp"
 #include "data_systems/lazy_load/lazy_load_system.hpp"
 #include "data_systems/offline.hpp"
-#include "evaluation/detail/evaluation_stack.hpp"
+#include "evaluation/evaluation_stack.hpp"
 
 #include "data_interfaces/system/idata_system.hpp"
 
@@ -170,7 +170,7 @@ AllFlagsState ClientImpl::AllFlagsState(Context const& context,
     AllFlagsStateBuilder builder{options};
 
     EventScope no_events;
-    evaluation::detail::EvaluationStack stack;
+    evaluation::EvaluationStack stack;
 
     auto all_flags = data_system_->AllFlags();
 
@@ -293,7 +293,7 @@ EvaluationDetail<Value> ClientImpl::VariationInternal(
                               event_scope, std::nullopt);
     }
 
-    evaluation::detail::EvaluationStack stack;
+    evaluation::EvaluationStack stack;
     EvaluationDetail<Value> result =
         evaluator_.Evaluate(*flag_rule->item, context, stack, event_scope);
     return PostEvaluation(key, context, default_value, result, event_scope,
